requirements.txt
flask==1.1.2
beautifulsoup4==4.6.3
requests==2.20.1
line-bot-sdk==1.19.0
datetime
matplotlib